requirements.txt
-i https: // pypi.org/simple
cachetools == 3.1.0
certifi == 2019.3.9
chardet == 3.0.4
click == 7.0
python_version != '3.3.*'
enum34 == 1.1.6
python_version < '3.4'

flask-login == 0.4.1
flask-wtf == 0.14.2
flask == 1.0.2
futures == 3.2.0
python_version < '3.2'
google-api-core == 1.9.0
google-auth == 1.6.3
google-cloud-core == 0.29.1
google-cloud-firestore == 0.32.1
google-cloud-storage == 1.15.0
google-resumable-media == 0.3.2
python_version >= '2.6'
googleapis-common-protos == 1.5.9
python_version >= '2.6'
grpcio == 1.20.0
idna == 2.8
python_version != '3.3.*'
itsdangerous == 1.1.0
python_version != '3.3.*'
jinja2 == 2.10.1
python_version != '3.3.*'
markupsafe == 1.1.1
python_version != '3.3.*'
marshmallow == 2.19.2
protobuf == 3.7.1
python_version >= '2.6'
pyasn1-modules == 0.2.5
pyasn1 == 0.4.5
pymongo == 3.8.0
pytz == 2019.1
requests == 2.21.0
python_version < '4'
rsa == 4.0
six == 1.12.0
python_version >= '2.6'
urllib3 == 1.24.2
python_version < '4'
werkzeug == 0.15.2
python_version != '3.3.*'
wtforms == 2.2.1
